pca -- Principal Component Analysis -- version 0
Integration of the pca development in Git.
The Doc directory contains pdf documentation on statistical analysis
The Data directory contains statistical data for testing
The gnuplot directory contains gnuplot configuation/plotting files
The output-data directory contains output of the program
The output-pict directory contains gnuplot graphs of the output data

al is the stub of the linear algebra library (matrix operations)
pca.c was the first version of pca, with static two-dimensional arrays
for matrices.